s2b-crawler/requirements.txt
playwright>=1.49
httpx>=0.28
python-dotenv>=1.0
